feat: rearrange components
